requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Pillow==10.0.1

gunicorn==21.2.0
rarfile==4.0
iyzipay==1.0.45

# Test dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2
requests==2.31.0 